CS 330/Software Design & Engineering
This is the software design & engineering portion of the ePortfolio, which is the enhanced version of my CS 330 final project. The chair and table files tie into the main project, room.cpp, to display a room with a single chair and table. The document is the narrative that coincides with this part of the project.