CS 340/Algorithms & Data Structure
These files are python files that are used in a MongoDB environment to display the ability to utilize CRUD functionality to create, read, update, and delete files within a database. This satisfies the algorithms & data structure component of the CS 499 ePortfolio project. The last file is the narrative that describes the process and enhancements.